* 508965 rounding up obligation calculation and changing type for Tonnage column and property (#183)

* 508965 - Rounds up obligation calculation result and changes tonne property type from float to integer

* 508965 - Rounding up obligation calculation result - migration script updated

* 508965 - Rounding up  - Added FC material back in

* 508965 - rounding up obligation calculation - Use Math.Ceiling rather than ...ToPositiveInfinity

* merging release 81 branch into main and pipeline changes (#186) (#187)

* syncing release 8 changes into 81 (#175)

* Exclude Prns that have already been updated when selecting Prns to send to NPWD (#171)



* Prevent NPWD Prns awaiting acceptance overwriting accepted and reject… (#173)

* Prevent NPWD Prns awaiting acceptance overwriting accepted and rejected Prns

---------



---------




* Map Prn material to matching value in Material table (#176) (#179)




* prod yaml file change missing from r81

---------


* Feature/509015 new endpoint getobligationcalculation (#182)

* Initial checkin

* Map Prn material to matching value in Material table (#176)


* Mark old methods as obselete

* Remove obselete tag

* remove old endpoint and update unit tests

* refactored repositories and unit tests

* merging release 81 branch into main and pipeline changes (#186)

* syncing release 8 changes into 81 (#175)

* Exclude Prns that have already been updated when selecting Prns to send to NPWD (#171)


* Prevent NPWD Prns awaiting acceptance overwriting accepted and reject… (#173)

* Prevent NPWD Prns awaiting acceptance overwriting accepted and rejected Prns

---------


---------


* Map Prn material to matching value in Material table (#176) (#179)


* prod yaml file change missing from r81

---------


* Consolidate obligation calculations for PC & FC

* Refactored to address PR comments

* Rolledback repository change

* PR comments updates

---------


* Feature/509015 new endpoint getobligationcalculation (#188)

* Initial checkin

* Map Prn material to matching value in Material table (#176)


* Mark old methods as obselete

* Remove obselete tag

* remove old endpoint and update unit tests

* refactored repositories and unit tests

* merging release 81 branch into main and pipeline changes (#186)

* syncing release 8 changes into 81 (#175)

* Exclude Prns that have already been updated when selecting Prns to send to NPWD (#171)


* Prevent NPWD Prns awaiting acceptance overwriting accepted and reject… (#173)

* Prevent NPWD Prns awaiting acceptance overwriting accepted and rejected Prns

---------


---------


* Map Prn material to matching value in Material table (#176) (#179)


* prod yaml file change missing from r81

---------


* Consolidate obligation calculations for PC & FC

* Refactored to address PR comments

* Rolledback repository change

* PR comments updates

* Fix PCFC grouping

* Rollback appsettings change

---------


* feature/509015 aggregate PC and FC values and enable multi organizations (#192)

* Initial checkin

* Map Prn material to matching value in Material table (#176)


* Mark old methods as obselete

* Remove obselete tag

* remove old endpoint and update unit tests

* refactored repositories and unit tests

* merging release 81 branch into main and pipeline changes (#186)

* syncing release 8 changes into 81 (#175)

* Exclude Prns that have already been updated when selecting Prns to send to NPWD (#171)


* Prevent NPWD Prns awaiting acceptance overwriting accepted and reject… (#173)

* Prevent NPWD Prns awaiting acceptance overwriting accepted and rejected Prns

---------


---------


* Map Prn material to matching value in Material table (#176) (#179)


* prod yaml file change missing from r81

---------


* Consolidate obligation calculations for PC & FC

* Refactored to address PR comments

* Rolledback repository change

* PR comments updates

* Fix PCFC grouping

* Rollback appsettings change

* Fix Aggregation of FC and PC

* updated unit test to improve coverage

* Added Unit test

* Improve coverage

* Add new unit test

* refactored pc fc consolidation

* Refactored aggregation and cleaned up unnecessary changes

* Remove unncessary changes

* undo codecoverage exclusions

* Rollback blank space changes

* Revert back model changes

* revert back blank space change

* fixed unit test

* Addressing SonarQube suggestions

---------


* Rounding up glass calculation to nearest tonne (#191)

* 508965 - Rounds up obligation calculation result and changes tonne property type from float to integer

* 508965 - Rounding up obligation calculation result - migration script updated

* 508965 - Rounding up  - Added FC material back in

* 508965 - rounding up obligation calculation - Use Math.Ceiling rather than ...ToPositiveInfinity

* 508965 - Rounding up glass

* 508965 - Rounding up glass calculation to nearest tonne - removed commented out lines

* 508965 - Rounding up obligation calculation - tonnage type from double to integer within classes

---------

+++ src/EPR.PRN.Backend.Obligation/Interfaces/IMaterialCalculationService.cs
@@ -2,7 +2,7 @@
 {
     public interface IMaterialCalculationService
     {
-        int Calculate(double target, double tonnage);
-        (int remelt, int remainder) CalculateGlass(double target, double remeltTarget, double tonnage);
+        int Calculate(double target, int tonnage);
+        (int remelt, int remainder) CalculateGlass(double target, double remeltTarget, int tonnage);
     }
 }
